TASK1
We created an end-to-end tic tac-toe Python project and implementing the knowledge of Python that we have gained throughout the  tutorial . Tic tac toe Python is a two-player game played on the 3X3 grid between two players. Each player chooses between X and O and the first player starts to draw the X on the space on the grid followed alternatively by the other until a player successfully marks a streak on the grid else if all spaces are filled the game is set to draw.